legacy/examples/distribute_graphsage/requirements.txt
pgl==1.1.0
pyyaml
paddlepaddle==1.6.1

scipy
redis==2.10.6
redis-py-cluster==1.3.6

